Move rest package to cmp package
